requirements.txt
Pillow==9.5.0
instagram_private_api==1.6.0
requests==2.28.2
bs4==0.0.1
python-dotenv==1.0.0
openai==0.27.4
